Создан модуль infoModule в store
