+++ test_project/src/App.vue
@@ -22,8 +22,7 @@
         mounted() {
             this.showToggle()
             this.$store.dispatch('productsModule/loadProducts');
-            this.$store.commit('basketModule/setBasket');
-            this.$store.dispatch('loadDescription')
+            this.$store.dispatch('infoModule/loadDescription')
         },
         methods:{
             showToggle(){
